requirements.txt
langchain_google_genai
langchain_openai
langchain_ollama
langchain~=0.3.9
langchain_chroma
langchain_community
rank_bm25
gradio
sqlmodel~=0.0.22
uvicorn~=0.32.1
pyjwt~=2.9.0
fastapi~=0.115.5
passlib~=1.7.4

loguru~=0.7.2
python-dotenv~=1.0.1
chromadb~=0.5.20
pydantic~=2.7.4
redis~=5.2.1
rq~=2.0.0
httpx~=0.27.2
requests~=2.32.3
starlette~=0.41.3
networkx~=3.4.2
nltk~=3.9.1
matplotlib~=3.9.2
llama_parse